fix sonar findings

+++ screwbox-core/src/main/java/de/suzufa/screwbox/core/Grid.java
@@ -62,8 +62,8 @@ public class Grid implements Serializable {
          * Distance doesn't consider {@link Grid#gridSize}.
          */
         public double distance(final Node other) {
-            final int deltaX = other.x - x;
-            final int deltaY = other.y - y;
+            final double deltaX = other.x - x;
+            final double deltaY = other.y - y;
             return Math.sqrt(deltaX * deltaX + deltaY * deltaY);
         }
     }
